# cmd/server/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This API handles Patients and Dentists.",
        "title": "Certified Tech Developer - Back End III",
        "termsOfService": "https://developers.ctd.com.ar/es_ar/terminos-y-condiciones",
        "contact": {
            "name": "API Support",
            "url": "https://developers.ctd.com.ar/support"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "paths": {
...
        "/turnos/byDNI": {
            "get": {
                "description": "Retrieve turnos by paciente's DNI",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Turno"
                ],
                "summary": "Get turnos by paciente's DNI",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Paciente DNI",
                        "name": "dni",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            }
        },
        "/turnos/create": {
            "post": {
                "description": "Create a new turno",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Turno"
                ],
                "summary": "Create a new turno",
                "parameters": [
                    {
                        "description": "Turno object",
                        "name": "turno",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain.Turno"
                        }
                    },
                    {
                        "type": "string",
                        "description": "TOKEN",
                        "name": "token",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/domain.Turno"
                        }
                    }
                }
            }
        },
        "/turnos/createByDniAndMatricula": {
            "post": {
                "description": "Create a new turno by Paciente's DNI and Odontologo's Matricula",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Turno"
                ],
                "summary": "Create a new turno by Paciente's DNI and Odontologo's Matricula",
                "parameters": [
                    {
                        "description": "Turno data",
                        "name": "turno",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/turnos.CreateTurnoData"
                        }
                    },
                    {
                        "type": "string",
                        "description": "TOKEN",
                        "name": "token",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/domain.Turno"
                        }
                    }
                }
            }
        },
        "/turnos/update/{id}": {
            "put": {
                "description": "Update an existing turno",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Turno"
                ],
                "summary": "Update an existing turno",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Turno ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Turno object",
                        "name": "turno",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain.Turno"
                        }
                    },
                    {
                        "type": "string",
                        "description": "TOKEN",
                        "name": "token",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Updated",
                        "schema": {
                            "$ref": "#/definitions/domain.Turno"
                        }
                    }
                }
            }
        },
        "/turnos/{id}": {
            "get": {
                "description": "Retrieve a turno by ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Turno"
                ],
                "summary": "Get a turno by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Turno ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/domain.Turno"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete a turno by ID",
                "tags": [
                    "Turno"
                ],
                "summary": "Delete a turno",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Turno ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "TOKEN",
                        "name": "token",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Turno deleted successfully",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/turnos/{id}/update-field": {
            "put": {
                "description": "Update a specific field of a turno by ID",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Turno"
                ],
                "summary": "Update a specific field of a turno",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Turno ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Authentication Token",
                        "name": "token",
                        "in": "query",
                        "required": true
                    },
                    {
                        "description": "Name of the field to update",
                        "name": "field_name",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "New value for the field",
                        "name": "value",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "type": "string",
                        "description": "TOKEN",
                        "name": "token",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Field updated successfully",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "domain.CustomTime": {
            "type": "object",
            "properties": {
                "time.Time": {
                    "type": "string"
                }
            }
        },
...
        "domain.Turno": {
            "type": "object",
            "properties": {
                "descripcion": {
                    "type": "string"
                },
                "fecha": {
                    "$ref": "#/definitions/domain.CustomTime"
                },
                "id": {
                    "type": "integer"
                },
                "id_odontologo": {
                    "type": "integer"
                },
                "id_paciente": {
                    "type": "integer"
                }
            }
        },
        "turnos.CreateTurnoData": {
            "type": "object",
            "properties": {
                "descripcion": {
                    "type": "string"
                },
                "dni": {
                    "type": "string"
                },
                "fecha": {
                    "$ref": "#/definitions/domain.CustomTime"
                },
                "matricula": {
                    "type": "string"
                }
            }
        }
    }
}